+++ backend/common/widgets.py
@@ -1,4 +1,7 @@
-from import_export.widgets import CharWidget
+from django.utils.encoding import smart_str
+
+from import_export.widgets import CharWidget, ManyToManyWidget
+from taggit.managers import _TaggableManager
 
 
 class UUIDWidget(CharWidget):
@@ -17,3 +20,27 @@ class UUIDWidget(CharWidget):
         if value in self.NULL_VALUES:
             return None
         return super().clean(value, row, *args, **kwargs)
+
+
+class TagsWidget(ManyToManyWidget):
+    """
+    Widget that handles Django Taggit 'tags' field.
+
+    :param separator: Defaults to ``','``.
+    """
+
+    def __init__(self, separator=None):
+        if separator is None:
+            separator = ","
+        self.separator = separator
+        super(ManyToManyWidget).__init__()
+
+    def clean(self, value, row=None, *args, **kwargs):
+        return value.split(self.separator) if value else []
+
+    def render(self, value, obj=None):
+        if isinstance(value, _TaggableManager):
+            tags = [smart_str(getattr(obj, "name")) for obj in value.all()]
+        else:
+            tags = [str(v) for v in value]
+        return self.separator.join(tags)